requirements.txt
python-bot-telegram 
fastapi
uvicorn[standard]